lots of small fixes related to the rythmic rodent
git-svn-id: svn://localhost/ardour2/branches/2.0-ongoing@2967 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ gtk2_ardour/time_axis_view.cc
@@ -1122,8 +1122,8 @@ TimeAxisView::show_temporary_lines (const vector<nframes64_t>& pos)
 	list<ArdourCanvas::SimpleLine*>::iterator l;
 
 	for (i = pos.begin(), l = temp_lines.begin(); i != pos.end() && l != temp_lines.end(); ++i, ++l) {
-		(*l)->property_x1() = editor.frame_to_pixel ((double) *i);
-		(*l)->property_x2() = editor.frame_to_pixel ((double) *i);
+		(*l)->property_x1() = editor.frame_to_pixel (*i);
+		(*l)->property_x2() = editor.frame_to_pixel (*i);
 	}
 }
